added genres to author content on login and reorganized login router
